+++ pywb/rewrite/test/test_rewrite_live.py
@@ -84,14 +84,14 @@ def test_example_domain_specific_3():
 
 
 def test_post():
-    buff = BytesIO('ABCDEF')
+    buff = BytesIO('ABC=DEF')
 
     env = {'REQUEST_METHOD': 'POST',
-           'HTTP_ORIGIN': 'http://example.com',
-           'HTTP_HOST': 'example.com',
+           'HTTP_ORIGIN': 'http://httpbin.org',
+           'HTTP_HOST': 'httpbin.org',
            'wsgi.input': buff}
 
-    status_headers, resp_buff = get_rewritten('http://example.com/', urlrewriter, env=env)
+    status_headers, resp_buff = get_rewritten('http://httpbin.org/post', urlrewriter, env=env)
     assert status_headers.get_statuscode() == '200', status_headers